change rating type from Int to Double #78

+++ src/main/kotlin/com/nazonazo_app/shit_forces/account/AccountInfo.kt
@@ -2,8 +2,8 @@ package com.nazonazo_app.shit_forces.account
 
 data class AccountInfo(
     val name: String,
-    val rating: Int,
-    val innerRating: Int,
+    val rating: Double,
+    val innerRating: Double,
     val partNum: Int,
     val passwordHash: String,
     private val auth: String
@@ -22,8 +22,8 @@ data class AccountRatingChangeHistory(
     val accountName: String,
     val contestName: String,
     val indexOfParticipation: Int,
-    val prevRating: Int,
-    val newRating: Int,
+    val prevRating: Double,
+    val newRating: Double,
     val performance: Int
 )
 data class RequestAccount(
